Celebrity_Classification/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.5
fonttools==4.55.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.7.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
